src/requirements.txt
flask~=2.2.2
flask_session
gunicorn
pyTelegramBotAPI
requests~=2.28.1
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
creds==1.0.15
cycler==0.11.0
DateTime==4.7
Flask==2.2.2
flask_session
gunicorn

fonttools==4.37.4
frozenlist==1.3.1
grpcio==1.49.1
grpcio-tools
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
multidict==6.0.2
numpy==1.23.4
packaging==21.3
pandas==1.5.0
pathlib==1.0.1
Pillow==9.2.0
protobuf==3.20.3
pydantic==1.10.2
pyparsing==3.0.9
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
six==1.16.0
ta==0.10.2
tinkoff==0.1.1
tinkoff-investments==0.2.0b39
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
zope.interface==5.5.0
termcolor~=2.2.0